backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
mysql-connector-python>=8.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
setuptools>=65.0.0
wheel>=0.38.0

